# fix: failed to select blocks for recluster