… src/query/service/tests/it/storages/fuse/operations/mutation/recluster_mutator.rs
… src/query/storages/fuse/src/table_functions/clustering_information/clustering_information.rs
… src/query/storages/fuse/src/table_functions/table_args.rs
… tests/sqllogictests/suites/base/09_fuse_engine/09_0008_fuse_optimize_table.test