I hereby agree to the terms of the CLA available at: https://docs.databend.com/dev/policies/cla/

## Summary

This Pull Request (PR) addresses and fixes a bug where the condition `min > max` could occur after sorting and then truncating fields. The specific changes are as follows:

1. **Bug Fix**: The original logic could result in a `min > max` situation when truncating fields after sorting, causing incorrect behavior.
2. **Improvement**: To prevent this issue, when the cluster key is of string type, a `substr` expression is introduced. This change ensures that the substring operation is applied before sorting, thereby avoiding the mentioned bug.

By implementing this change, the process ensures that when dealing with string type cluster keys, the sorting and truncating operations are performed in the correct order, preventing logical errors in data processing.

Before:
```sql
mysql> create table t14(a string, b string) cluster by(a,b);
Query OK, 0 rows affected (0.13 sec)

mysql> insert into t14 values('123456789', 'abc');
Query OK, 1 row affected (0.06 sec)

mysql> insert into t14 values('12345678', 'efg');
Query OK, 1 row affected (0.13 sec)

mysql> alter table t14 recluster;
Query OK, 2 rows affected (0.22 sec)

mysql> select * from clustering_information('default','t14');
+-------------+-------------------+----------------------+-------------------------+------------------+---------------+-----------------------+
| cluster_key | total_block_count | constant_block_count | unclustered_block_count | average_overlaps | average_depth | block_depth_histogram |
+-------------+-------------------+----------------------+-------------------------+------------------+---------------+-----------------------+
| (a, b)      |                 1 |                    0 |                       0 |              NaN |           NaN | {}                    |
+-------------+-------------------+----------------------+-------------------------+------------------+---------------+-----------------------+
1 row in set (0.10 sec)
Read 1 rows, 448.00 B in 0.071 sec., 14.09 rows/sec., 6.16 KiB/sec.

mysql> select * from t14;
+-----------+------+
| a         | b    |
+-----------+------+
| 12345678  | efg  |
| 123456789 | abc  |
+-----------+------+
2 rows in set (0.07 sec)
Read 2 rows, 73.00 B in 0.017 sec., 118.18 rows/sec., 4.21 KiB/sec.
```
After:
```sql
mysql> alter table t14 recluster;
Query OK, 2 rows affected (0.22 sec)

mysql> select * from clustering_information('default','t14');
+-------------+-------------------+----------------------+-------------------------+------------------+---------------+-----------------------+
| cluster_key | total_block_count | constant_block_count | unclustered_block_count | average_overlaps | average_depth | block_depth_histogram |
+-------------+-------------------+----------------------+-------------------------+------------------+---------------+-----------------------+
| (a, b)      |                 1 |                    0 |                       0 |              0.0 |           0.0 | {"00001":1}           |
+-------------+-------------------+----------------------+-------------------------+------------------+---------------+-----------------------+
1 row in set (0.10 sec)
Read 1 rows, 448.00 B in 0.071 sec., 14.09 rows/sec., 6.16 KiB/sec.

mysql> select * from t14;
+-----------+------+
| a         | b    |
+-----------+------+
| 123456789 | abc  |
| 12345678  | efg  |
+-----------+------+
2 rows in set (0.10 sec)
Read 2 rows, 73.00 B in 0.013 sec., 155.4 rows/sec., 5.54 KiB/sec.
```

- Fixes #[Link the issue here]

## Tests

- [ ] Unit Test
- [x] Logic Test
- [ ] Benchmark Test
- [ ] No Test  - _Explain why_

## Type of change

- [x] Bug Fix (non-breaking change which fixes an issue)
- [ ] New Feature (non-breaking change which adds functionality)
- [ ] Breaking Change (fix or feature that could cause existing functionality not to work as expected)
- [ ] Documentation Update
- [ ] Refactoring
- [ ] Performance Improvement
- [ ] Other (please describe):

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/datafuselabs/databend/15667)
<!-- Reviewable:end -->